Fix qos/test_pfc_pause.py for dualtor.
The test sends server to server traffic. It tries to retrieve lossless priorities as
well as DSCP_TO_TC_MAP asscoiated with the ports. For this it checks one interface
at random for these two parameters.

In t0 all ports have same lossless priorities as well as DSCP_TO_TC_MAP but this is
not true for dualtor as it has different qos config for uplink and downlink ports.

Proposed solution is to check a downlink port for DSCP_TO_TC_MAP and lossless
priorities. Also the testcases which run for PG2 and 6 should be skipped for dualtor
as no DSCP maps to them in case of downlink ports.

Another issue is that the randomly selected Tor should be made active in dualtor for
which the fixture `toggle_all_simulator_ports_to_rand_selected_tor_m` is used.