Add callback for thread to call after invoking ptf command (#11926)
## Description of PR

Summary:
Fixes # (issue) 27085944

### Type of change

- [x] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] Test case(new/improvement)

## Approach
#### What is the motivation for this PR?

In one test case failure, we saw that 30 second max timeout was not enough time for the on change update to come. Increasing it to 60 seconds, but we expect the notification to come <=30 seconds.

Improve thread communication via passing a callback function instead of passing a results array

#### How did you do it?

Code change

#### How did you verify/test it?

Manual test / pipeline

Signed-off-by: yenlu <yenlu@cisco.com>