[dhcp_relay] Validate DHCPv6 multicast packets can be trapped to CPU (#11978)
To make sure DHCPv6 multicast packets can be received by device, we add a test to send DHCPv6 multicast packet with specific transaction ID from server to device, and check if tcpdump can capture the packet.

What is the motivation for this PR?
We need to make sure DHCPv6 multicast packets can be received by device.

How did you do it?
We add a test to send DHCPv6 multicast packet with specific transaction ID from server to device, and check if tcpdump can capture the packet.

How did you verify/test it?
For mx topo, verified on Nokia-7215 testbed with 202305 image and 202205 image.

Signed-off-by: yenlu <yenlu@cisco.com>