Dynamic PORT selection for PFC storm testing (#10425)
* Dynamic PORT selection for PFC storm testing

In the test file : telemetry/events/swss_events.py
in the function generate_pfc_storm, in the line
queue_oid = duthost.get_queue_oid(PFC_STORM_TEST_PORT, PFC_STORM_TEST_QUEUE)
The PFC_STORM_TEST_PORT - the port chosen to test the PFC storm on - is assigned a fixed value of “Ethernet4”, whereas, the T0-crocodile DUT does not have a port Ethernet4 associated with it’s ASIC instance. Therefore this test is passing when assigning a value of Ethernet8 or Ethernet16 which are valid ports.

Changing fixed assigned value, replacing it with a dynamically selected port from the list of ports that have both open and admin status as "UP"

* Adding format filter to interface

the yang validation needs interfaces with a particular format of name, this change ensures that we choose a suitable interface

* Fixing import statement position

Signed-off-by: yenlu <yenlu@cisco.com>