raise Exception when recover failed (#12051)
Approach
What is the motivation for this PR?
When an error occurs during recovery, the previous script would simply log the error and return, which Elastictest interprets as a successful operation. However, in this PR, we have implemented a change where if the script encounters a failure, it will now raise an Exception instead.

How did you do it?
Raise an exception instead of simply log the error

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>