Merged PR 5109: Restrict the scope of retry IPv6 (#12069)
Restrict the scope of retry IPv6 (#12069)

Approach
What is the motivation for this PR?
Refine the IPv6 retry logic, only retry on certain IPv4 addr unavailable scenarios.

How did you do it?
Check the Exception type and the message.

How did you verify/test it?
Run on different labs, the it works well