Update gnmi test to support t1 backend (#12050)
## Description of PR
Summary:
Fix gnmi test for DUT which all interfaces are routed.
Fixes # (issue)
Microsoft ADO: 27150366

## Approach
#### What is the motivation for this PR?
GNMI test failed on DUT which all interfaces are routed.
Some DUT do not have portchannel and all the interface type are "routed".
I was using "routed" as key to detect portchannel, but this logic does not work for t1 backend topology.

#### How did you do it?
I use lanes to detect portchannel, if lanes is "N/A", this is portchannel or other unwanted port.

#### How did you verify/test it?
Run gnmi end to end test.

Signed-off-by: yenlu <yenlu@cisco.com>