Refine IPv6 conversion comment, enhance IPv6 test to ignore mux error (#12214)
Description of PR
Summary:
For dualtor duts, we set up mux simulator on the servers,
but if the server doesn't have IPv6 addr, the mux simulator is unavailable,
Then y cable issue is reported, since the IPv6 test only focus on the mgmt plane,
we can ignore this error log

Approach
What is the motivation for this PR?
Fix the error when dualtor testbed runs into the mgmt IPv6 test module.

How did you do it?
Ignore the error log.

How did you verify/test it?
Run on physical testbeds, and it passed.

Signed-off-by: yenlu <yenlu@cisco.com>