Merged PR 5285: Skip rotating password for KVM testbed
Skip rotating password for KVM testbed

PR checkers of the sonic-buildimage internal repository is using internal and
internal-xxxx branches of sonic-mgmt-int repo.

The previous PR for rotating DUT password was intended for physical testbeds.
For KVM testbed, it should be OK to keep using the default password.

This change added condition for the step of changing default password to apply
only for non-KVM DUT.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>