Update macsec tests to use JSON output of `ip macsec show`
The current approach of parsing string output is a bit fragile; if
iproute2 decides to change the formatting, or the ordering of
attributes, then parsing could break. Fortunately, most of iproute2's
commands support JSON output.

Switch to parsing JSON instead of using regex. This should make it more
robust.

This is needed for upgrading docker-syncd-vs to Bookworm.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>