[multi-password] Add action plugin to get the hash of currently used password (#12216)
What is the motivation for this PR?
The multi_password_ssh plugin enhanced the ansible ssh plugin to support trying multiple passwords.
In case we need to know which password is used, some enhancement is required.

How did you do it?
This change enhanced the multi_password_ssh plugin to store the hash of current password in its options.
A new action plugin "current_password" is added to run a command module and try to get the stored hash
of current password. The hash of current password will be added to the module's result.

How did you verify/test it?
Tested with playbook like below:
```
- hosts: sonic
  gather_facts: no
  tasks:
  - name: Run command
    action: current_password
    args:
      argv: ["whoami"]
```

Signed-off-by: yenlu <yenlu@cisco.com>