Refactor upgrade path tests (#11698)
* Refactor the upgrade path tests

Currently, the upgrade path tests (of which there are two defined) only
support a single warm upgrade. Also, there is duplicated code between
the two test definitions, and doesn't include some checks that ideally
should be done. This makes adding support for a double-upgrade test a
bit harder.

Refactor the existing code to move the duplicated code into a helper
function. Add support for running pre-upgrade and post-upgrade hooks
(which are already provided in the `AdvancedReboot` class). Add checks
for sonic-installer errors (which likely wouldn't happen with current
images, but it's probably nice-to-handle, just in case) and for S6100
devices potentially needing a soft-reboot instead of a normal reboot.
Also add support for optionally enabling CPA. Finally, add a test case
for testing double-upgrade.

This also makes it easier to add support for some other cases, such as
MAC jumping test case.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>