platform_tests: Use VariableManager for skip modules (#11242)
When using the TestbedProcessing script to generate the inventory
files, the `skip_modules` variable under the `devices` section is
ignored. We can instead use the `host_vars` section which adds the
variable to the `host_vars/${host}.yml` file, but this file is ignored
by InventoryManager, which the tests use to query the `skip_modules`
variable.

This can be fixed by using VariableManager instead, which considers
all sources of inventory, including the `host_vars` and `group_vars`
directories.

With this change, the following config works:

```
host_vars:
  cmp210:
    skip_modules:
      'line-cards': ['LINE-CARD3']
```

Signed-off-by: yenlu <yenlu@cisco.com>