Enhance the "switch" connection plugin to hand Passwordless Login (#12318)
What is the motivation for this PR?
Case failed with error message "msg": "Establish connection to server failed after tried 3 times."

How did you do it?
Ansible uses the option "'ControlPersist=60s'" for SSH connection, it would allow the master connection to remain open after the initial SSH session has been established. This option helps to speed up subsequent SSH connections to the same host by reusing the existing connection.
But in the current script, the code always waits for the '[Pp]assword:' prompt, causing a timeout in passwordless login.

Break if the prompt is " '>', '#'" to handle the passwordless logdin

How did you verify/test it?
Run the original case.

Signed-off-by: xuliping <xuliping@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>