Adjust the number of ipv6 nd in script `arp/test_stress_arp.py`. (#12197)
What is the motivation for this PR?
In script arp/test_stress_arp.py, we want to test the maximum number of entries that can be supported by the DUT, and to ensure that the addition and removal of these entries are handled correctly. In this PR, we adjust the code to consider both the available IPv6 neighbor entries and the FDB available entries, which will provide a more comprehensive understanding of the DUT's capabilities and limitations.

How did you do it?
Adjust the code to consider both the available IPv6 neighbor entries and the FDB available entries

Signed-off-by: yenlu <yenlu@cisco.com>