Force root fanout task to use connection "switch" (#12435)
The task for changing root fanout port vlan must be executed over "switch"
connection instead of other ansible connection plugins.

However, when we set "ansible_connection" variable to "multi_passwd_ssh"
for lab server, somehow the "connection: switch" for the "Change root fanout
port vlan" task is not working. Consequently this task is not executed over
the "switch" connection plugin and fails.

To workaround this issue, we can explicitly set "ansible_connection" variable
to "switch" in previous steps. This will force the subsequent tasks to use the
"switch" connection plugin.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>