[test_po_cleanup] Do a safe config reload (#12346)
For test_po_cleanup and test_po_cleanup_after_reload, do a safe reload
and wait for BGP to come up. This makes sure that the next test case
that runs gets the testbed in a good, stable state.

This is particularly needed on slow-CPU devices where a proper config
reload will take some time to run.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>