Unify console access to use account from group vars (#12429)
What is the motivation for this PR?
In this PR, we streamlined the process of retrieving console usernames by consolidating the source to a single location: the group vars. This change simplifies our codebase and enhances maintainability. The field 'proxy' in console link CSV under the ansible/files folder is no longer in use for this purpose.

How did you do it?
Streamlined the process of retrieving console usernames by consolidating the source to a single location: the group vars.

Signed-off-by: yenlu <yenlu@cisco.com>