Fix test_bgp_update_time by increasing threshold (#10989)
What is the motivation for this PR?
The change is to increase the check threshold, because our tests randomly failed with measurements like 2.1 and the original value looks to be arbitrary set in #7907

How did you do it?
Increase the threshold from 2 to 2.5

How did you verify/test it?
The test can reliably pass after the change

Any platform specific information?
Arista 7260

Signed-off-by: yenlu <yenlu@cisco.com>