Update component version index in fwutil common (#11955)
In fwutil tests, it will always revert the component version to the version at index 0,
since this index is expected to be the latest version.

This is based on the assumption that the firmware.json file is maintained with the same structure.
The JSON file will list 2 versions per FW component (i.e, ONIE/BIOS/CPLD),
The 2 versions are always listed in the same order which is:

The first version is the newer version
The second version is the older version

Summary:
1. Giving index 0 a meaningful name will help the user better understand the JSON structure and test flow
2. Help debug and avoid file errors
3. Better coding practice since we remove the use of magic numbers.

Signed-off-by: yenlu <yenlu@cisco.com>