Add 16 data plane test to T0 PR checker (#12466)
What is the motivation for this PR?
This is part of the shift left effort by adding more tests to PR checkers to have better coverage in earlier stage of development.

How did you do it?
Add 16 data plane test to T0 PR checker and sort from A to Z

Signed-off-by: yenlu <yenlu@cisco.com>