[macsec]: Add a sleep at the link flap retry (#12324)
Approach
What is the motivation for this PR?
There is a flaky failure on the macsec_test_fault_handling. It's due to the lag of DUT exceeding MKA_TIMEOUT that triggers a rekey.

How did you do it?
To mitigate this, retry the test after a while with a few seconds of idle time.

How did you verify/test it?
Check Azp

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>