Fix the module `test_sfpshow.py` and add it in PR test (#12505)
What is the motivation for this PR?
The module test_sfpshow.py will fail on kvm testbed because of the failure

  File "/var/src/sonic-mgmt/tests/platform_tests/sfp/util.py", line 36, in get_dev_conn
    dev_conn = conn_graph_facts["device_conn"][duthost.hostname]
KeyError: 'vlab-01'

In this PR, we fix this issue, and add it into PR test.

How did you do it?
Use method get to get the value from a python dict. And set it {} if there is no such key.

Signed-off-by: yenlu <yenlu@cisco.com>