Fix route/test_route_consistency.py
Issue: After announcing the routes, it is taking some time for kernel
routes (IP-in-IP tunnel routes from standby ToR to active ToR) to be
populated in ASIC_DB. And current sleep_interval is not enough for
kernel routes to be populated in some of the runs and thus test is
failing because of mismatch in pre_test_route_snapshot and
post_test_route_snapshot.

Fix: Increasing the sleep_interval (to ~100 secs) for time.sleep
after announcing the routes (announce or withdraw).