Add test_stress_arp.py into nightly (#11928)
What is the motivation for this PR?
Because of the issue #7874, we skip script arp/test_stress_arp.py. In PR #11922 , we fix this issue, so add this script into nightly test again.

How did you do it?
Remove the condition in tests_mark_conditions.yaml and add the pytest mark.
This test is only for topology t0, because we need to get VLAN_INTERFACE in config, which is available for t0.