Add device type mark in some modules.  (#12496)
What is the motivation for this PR?
In this PR, I add mark pytest.mark.device_type('physical') in some modules, which means the module can only be ran on physical testbeds.

The reasons are below:
* generic_config_updater/test_pfcwd_status.py
* PFCWD needs some counters, and these counters don't support kvm testbeds.
* generic_config_updater/test_pg_headroom_update.py
  In this test, it needs to execute command sonic-db-cli CONFIG_DB keys *BUFFER_PROFILE\\|pg_lossless* on DUT and gets the return value, but there the table BUFFER_PROFILE doesn't exist on kvm, thus this command returns None.
* platform_tests/api
  All the cases under platform_tests/api don't support kvm testbeds
* platform_tests/daemon
  The features to be tested under this folder are not enabled in x86_64-kvm_x86_64-r0.

How did you do it?
Add the mark pytest.mark.device_type('physical') in the module which only support physical testbed.

Signed-off-by: yenlu <yenlu@cisco.com>