[202205][Container checker] Check if the container is in running state before stopping it (#12155)
Cherry Pick PR #11397 to 2205

Refer: Issue #10475
In the resolution process, we verify the status of the container under test to ensure it is running. If the container is found to be in exited stated, we initiate a config_reload to guarantee its running state before proceeding with the test case. To maintain accurate logging post config_reload, an additional 10-second delay is introduced for the log analyzer, compensating for the reset of the monit rc timer.

In certain scenarios, the auto-restart feature may inadvertently get enabled due to other test cases, which contradicts the prerequisite of having it disabled for the current test case. Consequently, a corrective measure has been implemented to ensure that auto-restart is appropriately disabled as part of the fix.

What is the motivation for this PR?
Issue: #10475

How did you do it?
If the container in test is not running -> Perform config_reload to ensure its running state
Sleep time is increased for 10s in case of config_reload for Log_Analyzer
If the auto_restart feature is enabled -> Perform disable_container_autorestart to ensure the auto restart is disabled

How did you verify/test it?
Ran the test cases against a multi-asic line card in a T2 chassis.