Fix drop_packets/test_configurable_drop_counters.py (#12500)
Testcases fail in drop_packets/test_configurable_drop_counters.py fail in dualtor-aa because some of the traffic goes to the unselected ToR.

How did you do it?
Similar approach is taken as in #11921

How did you verify/test it?
Verified on Arista 7260 and Arista 7050 with both dualtor and dualtor-aa topology.

Signed-off-by: yenlu <yenlu@cisco.com>