Add IPv6-only test case for NTP (#12258) (#12580)
- [x] Test case(new/improvement)

Add support for testing NTP over IPv6. Additionally, test NTP when running with an IPv6-only network.

Also, this fixes the IPv6 address assignment for vlab-02 and makes sure it's given an address on the same subnet as the PTF container.

Tested on KVM and physical DUT with IPv6-only mgmt address.