Merged PR 5698: Update deep clean trigger pipeline not to use PAT
PAT life time will be reduced to at most 7 days. To avoid the effort required for frequent PAT rotation, we need to avoid using PAT in our pipelines.

This change updated yaml file of the deep clean trigger not to use PAT.

The deep clean pipeline was broken due to introduction of testbed health monitoring and recovery. This change does not fix that broken issue. This change only deprecate PAT. Need extra PR to fix the broken deep clean pipeline.