Reuse already-installed images when doing an upgrade
For `test_upgrade_path.py`, if the from image or the to image is already
installed, then instead of removing it and reinstalling it, reuse the
existing image by restoring it to the pre-first-boot state (so that when
it's booted into, it'll be like it's the first time). This saves some
time in downloading and installing the image file to this device.

For this to work, the image version needs to be passed in as a
parameter; otherwise, the version is not known, and the existing
behavior of removing and reinstalling the image will apply.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>