Merged PR 5745: Support skip locking testbed in sonic-metadata test
This PR is to add an option `SKIP_LOCKING_TESTBED` in sonic-metadata pipeline.

The value is `false` by default. Locking testbed will be skipped if the checkbox is checked.

![image.png](https://dev.azure.com/mssonic/12b9cbf4-b1d3-4768-8e49-669345c32e5d/_apis/git/repositories/5380e8f7-6e2a-4154-8dee-f3be7b096894/pullRequests/5745/attachments/image.png)
The change is required because we need to reserve the testbed for a long time for running multiple sonic-metadata test. And if the testbed is released, it will be taken by nightly test.