Add fixture to detect NTP issue for gnmi testing (#12584)
What is the motivation for this PR?
We received a lots of tickets for gnmi end to end test failure, and the root cause is NTP issue and system time is wrong.

How did you do it?
Add fixture to check DUT time, and then we don't need to spend time to identify NTP issue.

How did you verify/test it?
Run GNMI end to end test.

Signed-off-by: yenlu <yenlu@cisco.com>