Add a separate non-blocking PR check for onboarding new tests (#12729)
What is the motivation for this PR?
We are going to onboard a batch of test cases into PR test recently. To achieve this without disrupting the existing PR test processes, we add a separate non-blocking PR checker for onboarding new tests.

How did you do it?
Add a separate non-blocking PR checker for onboarding new tests

Signed-off-by: yenlu <yenlu@cisco.com>