Fix _backup_and_restore_config_db to properly restore config_db.json after upgrade (#12160)
What is the motivation for this PR?
test_warm_upgrade_sad_path fails in teardown because it is unable to restore config_db. This is because in the event of an upgrade, config_db backup is deleted if stored in /etc/sonic/.

How did you do it?
Instead of storing config_db backup in /etc/sonic/, store in /host instead.

How did you verify/test it?
Ran test_warm_upgrade_sad_path and verified it passes

Signed-off-by: yenlu <yenlu@cisco.com>