Remove module autorestart/test_container_autorestart.py in PR test (#12827)
What is the motivation for this PR?
In PR #12387, we add the module autorestart/test_container_autorestart.py into PR test. Although it passed in master, it will fail in internal due to the container restapi. In this PR, we comment this module temporarily to unblock internal branch code sync.

Signed-off-by: yenlu <yenlu@cisco.com>