Change the mark in module `test_link_flap.py` (#12504)
What is the motivation for this PR?
The module test_link_flap.py can support physical testbeds only， and it uses pytest.mark.platform('physical') before to identify. But usually, we use mark pytest.mark.device_type('physical') to identify, and in Elastictest, we will also use the mark device_type to collect and filter cases. So in this PR, we change this mark.

How did you do it?
Change the mark pytest.mark.platform('physical') to pytest.mark.device_type('physical')