Fix route/test_route_consistency.py (#12512)
What is the motivation for this PR?
After announcing the routes, it is taking some time for kernel routes (IP-in-IP tunnel routes from standby ToR to active ToR) to be populated in ASIC_DB. And current sleep_interval is not enough for kernel routes to be populated in some of the runs and thus test is failing because of mismatch in pre_test_route_snapshot and post_test_route_snapshot (lead to AssertionError.

How did you do it?
Increased the sleep_interval (to ~100 secs) for time.sleep after announcing the routes (announce or withdraw) to give more time for kernel routes to be populated in ASIC_DB.

How did you verify/test it?
Test is consistently passing with the fix (verified on Arista-7260CX3-C64).

Signed-off-by: yenlu <yenlu@cisco.com>