Update the pfcwd test cases according to the new implement of the pfcwd. (#12733)
Update the pfcwd test cases according to the new implement of the pfcwd
After the feature is completed, it criteria to trigger the pfcwd is much more strict than previous on Nvidia platform, So adjust the test case to properly trigger the pfcwd. The solution is to send background traffic during the pfc frame is sent from the fanout.