Fix the Keyerror issue caused by none `conn_graph_fact` in VS set-up (#12751)
What is the motivation for this PR?
To adapt to kvm testbed, we need to fix the issue -- conn_graph_fact is none in VS set-up.
In this PR, we use method get of a dict to get the value instead of [] to avoid key error when getting value from an empty dict conn_graph_fact.

Also, in this PR, after fixing this issue, we can add three platform tests into PR test.

Signed-off-by: yenlu <yenlu@cisco.com>