Fix sonic-mgmt testbed-cli interface_to_vms undefined failure. (#12890)
What is the motivation for this PR?
This change is to fix the error above, when VM list is empty.

How did you do it?
The fix is straightforward - to give it a default value that is an empty list when enumerating.

How did you verify/test it?
The change is run locally with topology with empty VMs, and it works.

Signed-off-by: yenlu <yenlu@cisco.com>