Fix add-topo failure when deploying topology with empty VM list. (#12932)
What is the motivation for this PR?
There are 2 issues are hitting, when deploying any topology without the VM list:

Ansible Bind topology to VMs task task will fail, because the vm list is empty, but it is checking the VM base in the VM list.
The generated minigraph will fail to load, because minigraph.py is not expecting any ACL interface with empty attached to list.
How did you do it?
This change fixes both issue with adding the empty list checks in these 2 places.

How did you verify/test it?
Run the add-topo and deploy-mg locally, both works.

Any platform specific information?
No.

Signed-off-by: yenlu <yenlu@cisco.com>