This test was enabled on T0. This test assigns an ip to Eth0 and (#12952)
checks if MAC resolution happens properly or not.
The same functionality is tested by test_neighbor_mac_noptf.py
This test fails on T0s because on T0s the Eth0 may b e part of a
port channel. As a result the test fails to assign an IP to the Interface.
Disablign this test on T0 is the right approach as the same functionality
is bening tested on T0 by test_neighbor_mac_noptf.py

Signed-off-by: yenlu <yenlu@cisco.com>