Include advanced-reboot ptf exception in pytest assertion (#12941)
What is the motivation for this PR?
Surface more information in our test reporting infrastructure to aid in debugging

How did you do it?
Added more information to a pytest assertion.

How did you verify/test it?
Artificially added an exception to the try-except block and saw that it was printed out in the pytest assertions.

Signed-off-by: yenlu <yenlu@cisco.com>