added Nokia-7215-A1 support for fdb testcases (#12807)
What is the motivation for this PR?
This PR helps add support for Nokia-7215-A1 in test_fdb.py testcases. Currently it is skipped as there is no support.

How did you do it?
Added arm64-nokia_ixs7215_52xb-r0 in setup_po2vlan in portchannel_to_vlan.py.

How did you verify/test it?
Ran the entire fdb suite to see it is not skipped.

Signed-off-by: yenlu <yenlu@cisco.com>