Make SSH session generator support generating session files for all devices, like fanout switches. (#12974)
What is the motivation for this PR?
Currently, the SSH repo generator will only generate the session files for DUT and PTF in testbeds, but not all the other devices, such as fanout switches and servers.

How did you do it?
This change updates the ssh_session_gen.py to parse the device inventory csv files, so we can generate all sessions for it.

How did you verify/test it?
Run it locally, and it outputs the session files as below. Besides the testbeds, now we also outputs the devices folder:

Signed-off-by: yenlu <yenlu@cisco.com>