sonic-mgmt: add th5 asic type for qos tests
See aristanetworks/sonic-qual.msft/issues/137
(ASIC type is incorrectly identified as jr2).

In conjunction with the addition of the
Quicksilver HW SKUs (occurring in a separate
pull request) the TH5 ASIC type is being
added to the qos base test file here so it
can be properly identified during testing,
such as in the case of test_qos_sai.py.