Optimize loganalyzer regex matching by using match() instead of findall() (#12956)
* Optimize loganalyzer regex matching by using match() instead of findall()

Loganalyzer takes an unusally long time to test syslog messages to see
if they match some expected regex. This is because of using `findall`
(search anywhere in the string for all instances) instead of `match`
(search from the starting of the string). The time it takes to search
for different starting positions of a match appears to be expensive.

Most of our match strings have a `.*` at the beginning anyways, and
we're not returning the exact matching substrings (only whether or not
the regex matches), so change to using `match`.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>