Don't call expensive fixtures if skipping sanity
If the `--skip_sanity` option has been given, then don't call expensive
fixtures and get the values of those fixtures just to end up not using
them. Instead, have a small function that only gets the request
fixture, and if sanity checks are not skipped, then call the full
function that gets all fixtures.

This saves about 20-30 seconds of time.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>