Enhance test scripts validation and traceability in PR checker  (#13079)
What is the motivation for this PR?
To ensure that a test script is included in the PR checker for the corresponding topology type,
this script will perform a validation check. Additionally, the return value will be enhanced to include more detailed information, such as the category of the test. Post-execution, the script will also append the scan time and track ID to the results. Finally, we will upload the result to Kusto.

How did you do it?
Get all the scripts under the folder tests and filter out the line pytest.mark.topology in the script.
The values in this mark are considered as the support topology type of this test script.
Then, checking if this test script is in the PR checker of corresponding topology type.