[pc]: Skip if there's not enough members in the VLAN (#13211)
What is the motivation for this PR?
For the t0-8-lag topology, there are only 15 interfaces in the VLAN that will be up. This means that for topos where we want to test 16-member LAGs, this will be impossible.

How did you do it?
As a general fix, if there's not enough members in the VLAN, skip the test case. This was already being done in one place, but that line of code just needed to be copied to another fixture.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>