Exclude scripts under "onboarding_" while counting PR checker coverage (#13270)
What is the motivation for this PR?
In PR #13079, we add a script to count how many scripts are included in PR checker. But the scripts in sets "onboarding_" should not be included, as this set is just for test, the scripts in these sets may not success. After the scripts in sets "onboarding_" are stable, we will move them to corresponding topology set.