Merged PR 6069: Add skipped test scripts in PR checker caculation.
In PR [#13079](https://github.com/sonic-net/sonic-mgmt/pull/13079), we enhance test scripts validation and traceability in PR checker. There are some skipped test scripts exclued in PR checker which should also be considered as "included in PR checker".  So in this PR, we add a seperate script to record these skipped scripts and caculate them in our analyse script.