Merged PR 5864: Accelerate transfer of metadata scripts
Accelerate transfer of metadata scripts

Ansible's copy module is very slow if there's many (hundreds or more)
files to copy. In the case of sonic-metadata, it takes 3-6 minutes to
copy the couple hundred files (220 files based on one sample). This is a
known issue, and is documented on the help page for this module.

The suggestion to fix this is to use the synchronize module instead,
which uses rsync to transfer the files. However, we don't have rsync
installed on our devices.

Instead, use the very basic tar/untar approach, where we tar up the
sonic-metadata scripts and extract it on the device. This takes about 7
seconds (based on local testing to a DX010 device).

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>