Fix arp/test_stress_arp.py to not try and populate 100% of available resources
Previously the test retrieves the maximum available space for fdb entries and ipv4 neighbours. It then uses the minimum of that as the amount of entries to add and attempts to add them via an ARP reply from the PTF container.

These tables utilize hashing to store the values, and maximizing it will cause hash collisions to happen. This fix will make the test attempt using 2/3 of the available remaining entry space instead of using all of it.