[Mellanox] Stabilize test_crm_route  (#12882)
* Fix test_crm_route issue: https://redmine.mellanox.com/issues/3771757

For mellanox device, the crm available counter is related to LPM tree. When shutdown all interfaces in vlan (e.g. vlan 1000),
it will cause the route (e.g. 192.168.0.1/21 )for vlan to be removed, we have only one route for the prefix (21), after it is removed, the corresponding prefix in LPM tree will be removed too, which will lead to the LPM tree structure is changed. LPM tree change might cause available counter change dramatically, and we cannot estimate how long the change is ready.
Therefore, it will lead the first case of test_crm_route fail occasionally because the expected available counter is not decreased. So, we add another route with the same prefix(21) as the vlan's so that the LPM tree is not changed.

Change-Id: I563b62d47e55b76965b84e9cc2a108b3134bc8ed