Merged PR 6138: Fix issue of getting access_token for lock_release.py
Fix issue of getting access_token for lock_release.py

The secret value used by lock_release was not stored in variable group as "secret". So it was exported to environment variable and can be directly used by the lock_release.py script. After stored it as "secret", the value needs to be explicitly passed down to lock_release.py as environment variable. Other variables in variable group are not secret and still can be directly used in lock_release.py. That's why only 1 variable needs to be passed down.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>