Wrap test_plan.py in AzureCLI task (#13435)
What is the motivation for this PR?
This is the fist step for eliminating client secrets. Purpose of this change is to allow test_plan.py to get access_token via service connection for accessing scheduler.

How did you do it?
This change wrapped all test_plan.py call in AzureCLI task instead of regular Bash task.

The next step is to update test_plan.py to get access_token via service connection.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>