Improve the step for installing azure-cli in template (#13448)
What is the motivation for this PR?
On some agents, the azure cli package is already installed, it's unnecessary to try installing it again.

How did you do it?
This change improved the step for installing azure cli to check if the az command is available. If no, try to install azure-cli. If yes, skip this step.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>