sonic-mgmt: Fix arp/test_unknown_mac by disabling ipv6
The DUT will occasionally (~5 mins) send a echo request through IPV6 to the ptf container (from `arp_update`), and this is casuing the fdb table to be populated - which is bad because the test expects the fdb table to be empty.

Temporarily disabling arp_update during the test will eliminate the problem.