Add support to test_sensors for Dynamic changes of PSUs (#12519)
What is the motivation for this PR?
Allow the sensors checks of sensors related to the PSU to be dependent on the PSUs installed on the DUT and not the platform.

How did you do it?
Created a static mapping between each PSU model the feature supports to the sensors of that PSU (generalized to be independent on the slot number the psu is installed in).

How did you verify/test it?
Ran the test on a platform with a new PSU to verify it checks the correct sensors.

Signed-off-by: yenlu <yenlu@cisco.com>