Fix flakiness of qos/test_qos_dscp_mapping.py
In some of the cases, after sending packets (2000) the queue counter
value is not reflecting correct value (counter value is less than
expected) and thus lead to test failure. In the issue state, reading the
counter value (in breakpoint) again shows the correct value. So the fix
is to increase the wait time (based on "counterpoll show")