[Snappi] - fixing swss service restart logic for basic PFCWD testcases in multi-asic environment (#13483)
The existing PFCWD basic testcases have restart SWSS service testcases.

However, these testcases do not address the multi-asic environment. The pull-request addresses this.

Summary:
Fixes # (issue)
The fix finds out the DUT and interfaces used in the test and based on that selects swss@0 or swss@1 to restart. The fix first verifies if the platform field exists and if yes, is set to 'broadcom-dnx' field. If this is not true, then code executes the default testcase.

Signed-off-by: yenlu <yenlu@cisco.com>